+++ app/src/main/java/com/xmartlabs/scasas/doapp/ApplicationComponent.java
@@ -5,7 +5,6 @@ import com.xmartlabs.scasas.doapp.controller.ServiceController;
 import com.xmartlabs.scasas.doapp.controller.SessionController;
 import com.xmartlabs.scasas.doapp.helper.DatabaseHelper;
 import com.xmartlabs.scasas.doapp.helper.GeneralErrorHelper;
-import com.xmartlabs.scasas.doapp.model.Group;
 import com.xmartlabs.scasas.doapp.module.AndroidModule;
 import com.xmartlabs.scasas.doapp.module.ControllerModule;
 import com.xmartlabs.scasas.doapp.module.GeneralErrorHelperModule;
@@ -18,16 +17,16 @@ import com.xmartlabs.scasas.doapp.ui.BaseActivity;
 import com.xmartlabs.scasas.doapp.ui.BaseAppCompatActivity;
 import com.xmartlabs.scasas.doapp.ui.BaseFragment;
 import com.xmartlabs.scasas.doapp.ui.FragmentWithDrawer;
-import com.xmartlabs.scasas.doapp.ui.GroupsListActivity;
-import com.xmartlabs.scasas.doapp.ui.GroupsListFragment;
-import com.xmartlabs.scasas.doapp.ui.SignInActivity;
-import com.xmartlabs.scasas.doapp.ui.SignInFragment;
+import com.xmartlabs.scasas.doapp.ui.grouplist.GroupsListActivity;
+import com.xmartlabs.scasas.doapp.ui.grouplist.GroupsListFragment;
+import com.xmartlabs.scasas.doapp.ui.signin.SignInActivity;
+import com.xmartlabs.scasas.doapp.ui.signin.SignInFragment;
 import com.xmartlabs.scasas.doapp.ui.MainActivity;
-import com.xmartlabs.scasas.doapp.ui.SignUpActivity;
-import com.xmartlabs.scasas.doapp.ui.SignUpFragment;
+import com.xmartlabs.scasas.doapp.ui.signup.SignUpActivity;
+import com.xmartlabs.scasas.doapp.ui.signup.SignUpFragment;
 import com.xmartlabs.scasas.doapp.ui.StartActivity;
-import com.xmartlabs.scasas.doapp.ui.TasksListActivity;
-import com.xmartlabs.scasas.doapp.ui.TasksListFragment;
+import com.xmartlabs.scasas.doapp.ui.tasklist.TasksListActivity;
+import com.xmartlabs.scasas.doapp.ui.tasklist.TasksListFragment;
 import com.xmartlabs.scasas.doapp.ui.ValidatableFragment;
 import com.xmartlabs.scasas.doapp.ui.WelcomeActivity;
 import com.xmartlabs.scasas.doapp.ui.WelcomeFragment;
@@ -62,23 +61,23 @@ public interface ApplicationComponent {
   void inject(BaseAppCompatActivity baseAppCompatActivity);
   void inject(SingleFragmentActivity singleFragmentActivity);
 
+  void inject(GroupsListActivity groupsListActivity);
   void inject(MainActivity mainActivity);
   void inject(SignInActivity signInActivity);
   void inject(SignUpActivity signUpActivity);
   void inject(StartActivity startActivity);
   void inject(TasksListActivity tasksListActivity);
   void inject(WelcomeActivity welcomeActivity);
-  void inject(GroupsListActivity groupsListActivity);
 
   void inject(BaseFragment baseFragment);
   void inject(FragmentWithDrawer fragmentWithDrawer);
   void inject(ValidatableFragment validatableFragment);
 
+  void inject(GroupsListFragment groupsListFragment);
   void inject(SignInFragment signInFragment);
   void inject(SignUpFragment signUpFragment);
   void inject(TasksListFragment tasksListFragment);
   void inject(WelcomeFragment welcomeFragment);
-  void inject(GroupsListFragment groupsListFragment);
 
   void inject(Controller controller);
   void inject(ServiceController serviceController);
