+++ app/src/main/java/com/xmartlabs/scasas/doapp/ui/grouplist/GroupsListActivity.java
@@ -1,10 +1,12 @@
-package com.xmartlabs.scasas.doapp.ui;
+package com.xmartlabs.scasas.doapp.ui.grouplist;
 
 import android.support.annotation.NonNull;
 
 import com.f2prateek.dart.HensonNavigable;
 import com.f2prateek.dart.InjectExtra;
 import com.xmartlabs.scasas.doapp.model.User;
+import com.xmartlabs.scasas.doapp.ui.BaseFragment;
+import com.xmartlabs.scasas.doapp.ui.SingleFragmentActivity;
 
 /**
  * Created by scasas on 3/20/17.
