+++ app/src/main/java/com/xmartlabs/scasas/doapp/ui/signup/SignUpActivity.java
@@ -1,8 +1,10 @@
-package com.xmartlabs.scasas.doapp.ui;
+package com.xmartlabs.scasas.doapp.ui.signup;
 
 import android.support.annotation.NonNull;
 
 import com.f2prateek.dart.HensonNavigable;
+import com.xmartlabs.scasas.doapp.ui.BaseFragment;
+import com.xmartlabs.scasas.doapp.ui.SingleFragmentActivity;
 
 /**
  * Created by santiago on 3/9/17.
